backend/requirements.txt
fastapi
uvicorn
groq
python-multipart
PyPDF2
python-docx
pydantic
python-dotenv
sqlalchemy
